big update, read desc
you can check the updates below or see a more in-depth summary at the end.

big stuff:
- made a new api made which i then made the main mode and made the old mode the api mode )
- DARK MODE (using prefers colour scheme)

not so big stuff:
- made default err msg more tidy
- tidy up code and old debug stuff
- made error handling a little better
- changed all branding to twittlet to match the github repo
- added twitter waifu icon >_<
- different button text for different modes (why am i listing this ... )

so originally i was making a "api mode" which was not really api related, it was just a cleaner, better way of fetching the data without being redirected to the api. but once i finished that i made the old mode the api mode and made the new api mode the old mode/ main mode.

standard mode; grabs the four data types which the twitter api returns; three different sized mp4 files and a .m3u8 file. at the moment i havent added a handling method that displays the file sizes and/ or removes the m3u8 file from the listing (bc i its pretty useless). will be a target soon though probably.

api mode: just a gui to direct you to the api, kinda useless but there none-the-less.